rename cache_change
