+++ dds/src/implementation/rtps/mod.rs
@@ -1,4 +1,5 @@
 pub mod behavior_types;
+pub mod cache_change;
 pub mod discovery_types;
 pub mod endpoint;
 pub mod entity;
@@ -8,7 +9,6 @@ pub mod message_receiver;
 pub mod messages;
 pub mod participant;
 pub mod reader;
-pub mod reader_history_cache;
 pub mod reader_locator;
 pub mod reader_proxy;
 pub mod types;
